+++ src/resources/views/theme-test/pages/tables.blade.php
@@ -1,0 +1,9 @@
+@extends($theme->layout('private'))
+
+@section('content')
+    <h1>Tables</h1>
+    <hr>
+
+    {!! widget('appshell::user.index.table')->render(Konekt\AppShell\Models\User::take(10)->get()) !!}
+
+@endsection